+++ src/Core/Dfe.Complete.Application/Projects/Models/ListProjectsGroupsModel.cs
@@ -1,0 +1,3 @@
+namespace Dfe.Complete.Application.Projects.Models;
+
+public record ListProjectsGroupsModel(string GroupId, string GroupName, string GroupIdentifier, string TrustUkprn, string IncludedEstablishments);